requirements.txt
sagemaker==2.90.0
boto3
scikit-learn
pandas
joblib
protobuf==3.20.*
PyYAML
pytest
xgboost
requests